fix: delete downloaded IPSW after extraction (#42)
* fix: delete downloaded IPSW after extraction

* eliminate the shell-pipe when mounting the system restore image

parse the volume_path in python instead of using grep and cut.

this prevents the program from continuing if the mount failed, and it also clarifies what fails: the mount or the parsing of its output.

Log the parse volume_path.

* adapt gcloud storage cp run to log output on failure rather than just raising an exception that contains the return code

this will be helpful in the future when the upload fails in the middle in we wonder what exit code 1 means exactly.